Render AccessToAddressBook or ReferralContacs in ReferFriends screen basde on hasAllowedToAccessContacts prop
