+++ src/locales/en/common.json
@@ -903,14 +903,6 @@
       "claimTokensError": "Error"
     },
     "instruction": {
-      "shareLink": {
-        "title": "Share your link",
-        "paragraph": "Invite your friends to join Pillar."
-      },
-      "giveSmartWallet": {
-        "title": "Give Smart Wallet for free",
-        "paragraph": "Friends who install Pillar with your link will get free Smart Wallet activation."
-      },
       "getTokens": {
         "title": "Get free PLR",
         "paragraph": "Earn meta-tokens for referring friends."
